+++ paper2html/paper.py
@@ -514,6 +514,10 @@ class Paper:
 
     @property
     def paragraphs(self):
+        """
+        段落のリスト．読む順序に合わせて塊ごとに並んでいる．
+        @return: list of list of PaperItem
+        """
         if not self._paragraphs:
             self._arrange_paragraphs()
         return self._paragraphs
@@ -535,40 +539,6 @@ class Paper:
                 body_separated = page.body_paragraphs[-1][-1].separated
 
             self._paragraphs.extend(page.footers)
-
-    def _paragraph2txt(self, paragraph):
-        """
-        段落内の改行を取り除く．pdfからのコピペの整形に使用していたころの名残．
-        """
-        result = "".join([item.text for item in paragraph])
-        patt = '([^(.|\n)])- ([^\n])'
-        result = re.sub(patt, r'\1\2', result)
-        patt = '([^(.|\n)])-\n([^\n])'
-        result = re.sub(patt, r'\1\2', result)
-        # patt = '\.\n'
-        # result = re.sub(patt, r'.\n\n', result)
-        patt = '([^(.|\n)])\n([^\n])'
-        result = re.sub(patt, r'\1 \2', result)
-        result = re.sub("([^\\.])\n", r"\1 ", result)
-        return result + "\n"
-
-    def _paragraph2img_elem(self, paragraph, i):
-        img_template = '<p id="img{}"><img alt="Figure" src="./{}" /></p>\n'
-        if paragraph[0].type == PaperItemType.Figure:
-            return img_template.format(i, os.path.relpath(paragraph[0].url, self.output_dir))
-        else:
-            return "\n".join([img_template.format(i, os.path.relpath(item.url, self.output_dir)) for item in paragraph])
-
-    def _paragraph2txt_elem(self, paragraph, i):
-        txt_template = '<p id="txt{}">{}</p>\n'
-        if len(paragraph) == 0:
-            return ""
-        if paragraph[0].type == PaperItemType.SectionHeader:
-            return '<h2 id="txt{}">{}</h2>\n'.format(i, self._paragraph2txt(paragraph))
-        elif paragraph[0].type == PaperItemType.Figure:
-            return txt_template.format(i, "")
-        else:
-            return txt_template.format(i, self._paragraph2txt(paragraph))
 
     def _draw_rect(self, bbox, ax, ec_str="#000000"):
         import matplotlib.patches as patches
@@ -597,276 +567,3 @@ class Paper:
             ax.set_aspect('equal')
             plt.savefig(pjoin(self.layout_dir, 'pdf2jpn%d.png' % page_n))
             plt.clf()
-
-
-class LocalHtmlPaper:
-    def __init__(self, paper, pdf_name):
-        self.paper = paper
-        self.pdf_name = pdf_name
-
-    def export(self):
-        css_content = '''
-        html, body {
-            height: 100%;
-            overflow: hidden;
-            margin: 0;
-        }
-        #split{
-            height: 100%;
-        }
-        #left {
-            float: left;
-            top: 0;
-            width: 50%;
-            height: 100%;
-            overflow: auto;
-            box-sizing: border-box;
-            z-index: 1;
-            padding: 50% 1.5em 50%;
-        }
-        #right{
-            float: left;
-            top: 0;
-            left: 50%;
-            width: 50%;
-            height: 100%;
-            overflow: auto;
-            box-sizing: border-box;
-            z-index: 2;
-            background-color: #FFFFFF;
-            padding: 50% 1.5em 50%;
-        }
-        '''
-        css_rel_path = pjoin('resources', 'stylesheet.css')
-        css_filename = pjoin(self.paper.output_dir, css_rel_path)
-        with open(css_filename, 'w', encoding="utf-8_sig") as f:
-            f.write(css_content)
-
-        top_html_template = '''
-            <!DOCTYPE html>
-            <html lang="en">
-              <head>
-                <meta http-equiv="Content-type" content="text/html;charset=utf-8" />
-                <link href="{}" rel="stylesheet" type="text/css" />
-                <title>
-                  {}
-                </title>
-              </head>
-              <body>
-                <div id="split">
-                  <header style="position: fixed;">
-                    <input type="button" value="30%" onclick="Zoom(0.33);"/>
-                    <input type="button" value="50%" onclick="Zoom(0.5);"/>
-                    <input type="button" value="65%" onclick="Zoom(0.65);"/>
-                    <input type="button" value="100%" onclick="Zoom(1);"/>
-                    <a href="{}">[Original PDF]</a>
-                  </header><br />
-                  <div id="left">
-                      {}
-                  </div>
-                  <div id="right">
-                      {}
-                  </div>
-                </div>
-                {}
-              </body>
-            </html>
-        '''
-        javascript = r'''
-        <script language="javascript" type="text/javascript">
-            const Zoom = function(rate) {
-                for (let i = 0; i < document.images.length; i++) {
-                    document.images[i].width = document.images[i].naturalWidth * rate;
-                    document.images[i].height = document.images[i].naturalHeight * rate;
-                }
-            }
-            const rightw = document.getElementById('right');
-            const leftw = document.getElementById('left');
-            const split = document.getElementById( 'split' );
-
-            var onscrollR = function() {
-             const top_ = split.scrollTop;
-             const bottom_ = top_ + split.clientHeight;
-             const center_ = (2/3) * top_ + (1/3) * bottom_;
-             for(var i = 0; i < rightw.children.length; i++) {
-              const txt_line = rightw.children[i];
-              const rect = txt_line.getBoundingClientRect();
-                if (rect.top <= center_ && center_ <= rect.bottom)
-                {
-                  const delta_rate = (center_ - rect.top) / rect.height;
-                  const img_line = document.getElementById(txt_line.id.replace('txt', 'img'));
-                  const delta_ = delta_rate * img_line.offsetHeight;
-                  leftw.scrollTo(0, img_line.offsetTop + delta_ - center_);
-                  break;
-                }
-              }
-            }
-            if( rightw.addEventListener )
-            {
-                rightw.addEventListener('scroll', onscrollR, false);
-            }
-        </script>
-        '''
-
-        def chunks(list, n):
-            if n == math.inf:
-                yield list
-                return
-            for i in range(0, len(list), n):
-                yield list[i:i + n]
-
-        html_pages = []
-        for paragraphs in chunks(self.paper.paragraphs, self.paper.n_div_paragraph):
-            img_content = "\n\n\n\n\n".join(
-                [self.paper._paragraph2img_elem(paragraph, i) for i, paragraph in enumerate(paragraphs)])
-            txt_content = "\n\n\n\n\n".join(
-                [self.paper._paragraph2txt_elem(paragraph, i) for i, paragraph in enumerate(paragraphs)])
-
-            html_pages.append([img_content, txt_content])
-
-        html_files = []
-        for i, page in enumerate(html_pages):
-            img_c, txt_c = page
-            output_filename = self.pdf_name + '_%d.html' % i
-            output_path = pjoin(self.paper.output_dir, output_filename)
-            with open(output_path, 'w', encoding="utf-8_sig") as f:
-                original_link = self.paper.output_dir + '.pdf'
-                f.write(top_html_template.format(css_rel_path, self.pdf_name, original_link, img_c, txt_c, javascript))
-            html_files.append(output_path)
-
-        return html_files
-
-
-class PatchedHtmlPaper(LocalHtmlPaper):
-    pass
-
-
-class CloseUppedHtmlPaper(LocalHtmlPaper):
-    pass
-
-
-def convert_paper_to_htmls(paper, pdf_name):
-    def chunks(list, n):
-        if n == math.inf:
-            yield list
-            return
-        for i in range(0, len(list), n):
-            yield list[i:i + n]
-    javascript = r'''
-    <script language="javascript" type="text/javascript">
-        const Zoom = function(rate) {
-            for (let i = 0; i < document.images.length; i++) {
-                document.images[i].width = document.images[i].naturalWidth * rate;
-                document.images[i].height = document.images[i].naturalHeight * rate;
-            }
-        }
-        const rightw = document.getElementById('right');
-        const leftw = document.getElementById('left');
-        const split = document.getElementById( 'split' );
-
-        var onscrollR = function() {
-         const top_ = split.scrollTop;
-         const bottom_ = top_ + split.clientHeight;
-         const center_ = (2/3) * top_ + (1/3) * bottom_;
-         for(var i = 0; i < rightw.children.length; i++) {
-          const txt_line = rightw.children[i];
-          const rect = txt_line.getBoundingClientRect();
-            if (rect.top <= center_ && center_ <= rect.bottom)
-            {
-              const delta_rate = (center_ - rect.top) / rect.height;
-              const img_line = document.getElementById(txt_line.id.replace('txt', 'img'));
-              const delta_ = delta_rate * img_line.offsetHeight;
-              leftw.scrollTo(0, img_line.offsetTop + delta_ - center_);
-              break;
-            }
-          }
-        }
-        if( rightw.addEventListener )
-        {
-            rightw.addEventListener('scroll', onscrollR, false);
-        }
-    </script>
-    '''
-    html_pages = []
-    for paragraphs in chunks(paper.paragraphs, paper.n_div_paragraph):
-        img_content = "\n\n\n\n\n".join([paper._paragraph2img_elem(paragraph, i) for i, paragraph in enumerate(paragraphs)])
-        txt_content = "\n\n\n\n\n".join([paper._paragraph2txt_elem(paragraph, i) for i, paragraph in enumerate(paragraphs)])
-
-        html_pages.append([img_content, txt_content])
-    css_content = '''
-    html, body {
-        height: 100%;
-        overflow: hidden;
-        margin: 0;
-    }
-    #split{
-        height: 100%;
-    }
-    #left {
-        float: left;
-        top: 0;
-        width: 50%;
-        height: 100%;
-        overflow: auto;
-        box-sizing: border-box;
-        z-index: 1;
-        padding: 50% 1.5em 50%;
-    }
-    #right{
-        float: left;
-        top: 0;
-        left: 50%;
-        width: 50%;
-        height: 100%;
-        overflow: auto;
-        box-sizing: border-box;
-        z-index: 2;
-        background-color: #FFFFFF;
-        padding: 50% 1.5em 50%;
-    }
-    '''
-    css_rel_path = pjoin('resources', 'stylesheet.css')
-    css_filename = pjoin(paper.output_dir, css_rel_path)
-    with open(css_filename, 'w', encoding="utf-8_sig") as f:
-        f.write(css_content)
-    html_files = []
-    for i, page in enumerate(html_pages):
-        top_html_template = '''
-            <!DOCTYPE html>
-            <html lang="en">
-              <head>
-                <meta http-equiv="Content-type" content="text/html;charset=utf-8" />
-                <link href="{}" rel="stylesheet" type="text/css" />
-                <title>
-                  {}
-                </title>
-              </head>
-              <body>
-                <div id="split">
-                  <header style="position: fixed;">
-                    <input type="button" value="30%" onclick="Zoom(0.33);"/>
-                    <input type="button" value="50%" onclick="Zoom(0.5);"/>
-                    <input type="button" value="65%" onclick="Zoom(0.65);"/>
-                    <input type="button" value="100%" onclick="Zoom(1);"/>
-                    <a href="{}">[Original PDF]</a>
-                  </header><br />
-                  <div id="left">
-                      {}
-                  </div>
-                  <div id="right">
-                      {}
-                  </div>
-                </div>
-                {}
-              </body>
-            </html>
-        '''
-        img_c, txt_c = page
-
-        output_filename = pdf_name + '_%d.html' % i
-        output_path = pjoin(paper.output_dir, output_filename)
-        with open(output_path, 'w', encoding="utf-8_sig") as f:
-            original_link = paper.output_dir + '.pdf'
-            f.write(top_html_template.format(css_rel_path, pdf_name, original_link, img_c, txt_c, javascript))
-        html_files.append(output_path)
-    return html_files
